Move header into rapids_logger subdirectory
